+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * replace `MLmetrics::MultiLogLoss()` with `ModelMetrics::mlogLoss()` in `logloss()` since it handles the case when the truth has a category with 0 counts but is represented in the probability matrix
 
+* add NPV and specificity to `evaluation()`
+
 # splendid 0.1.0
 
 * Default seed parameter value `NULL` does not invoke `set.seed()`